+++ integration-tests/cli/test/metadata.test.ts
@@ -1,7 +1,6 @@
 import test from 'ava';
-import path from 'node:path';
-import { exec } from 'node:child_process';
 import { differenceInMinutes } from 'date-fns';
+import path from 'node:path';
 
 import run, { clean } from '../src/run';
 import { extractLogs, getJSON } from '../src/util';
@@ -13,7 +12,15 @@ let lastCreated: Date;
 
 // For local dev, ensure the repo is clear
 test.before(async () => {
+  process.env.OPENFN_REPO_DIR = '/tmp/openfn/integration-tests/cli';
   await clean();
+});
+
+// Sleep after tests because weird exceptions keep popping up
+test.afterEach(async () => {
+  await new Promise((resolve) => {
+    setTimeout(resolve, 500);
+  });
 });
 
 // Generate metadata
@@ -27,7 +34,7 @@ test.serial(
 
     const logJson = extractLogs(stdout);
     const [outputPath] = logJson.at(-1).message;
-    t.regex(outputPath, /(repo\/meta\/)*(\.json)/);
+    t.regex(outputPath, /(repo\/.cli-cache\/)*(\.json)/);
 
     const metadata = getJSON(outputPath);
     t.is(metadata.name, 'test');
@@ -99,7 +106,7 @@ test.serial(
 
     const logJson = extractLogs(stdout);
     const [outputPath] = logJson.at(-1).message;
-    t.regex(outputPath, /(repo\/meta\/)*(\.json)/);
+    t.regex(outputPath, /(repo\/.cli-cache\/)*(\.json)/);
 
     const metadata = getJSON(outputPath);
     t.is(metadata.name, 'test');
@@ -127,16 +134,84 @@ test.serial('does not log credentials', async (t) => {
   t.notRegex(logString, new RegExp(sensitiveValue), 'i');
 });
 
-// This test should auto-install the common adaptor but then happily fail to generate metadata
+// When installing an adaptor that doesn't support metadata, the adaptor should be removed
 test.serial(
-  `openfn metadata -f -S "${state}" -a common --log-json --log info`,
+  `openfn metadata -S "${state}" -a openfn@3.0.0 --log-json --log info`,
   async (t) => {
     const { stdout } = await run(t.title);
 
     t.regex(stdout, /Generating metadata/);
     t.regex(stdout, /Installing packages.../);
-    t.regex(stdout, /Installed @openfn\/language-common@/);
+    t.regex(stdout, /Installed @openfn\/language-openfn@/);
     t.regex(stdout, /Installation complete in \d+\.\d+s/);
     t.regex(stdout, /No metadata helper found/);
+    t.regex(stdout, /Removing unsupported adaptor from disk/);
+    t.regex(stdout, /Adaptor removed and marked as unsupported/);
+    t.regex(stdout, /successfully removed @openfn\/language-openfn_/i);
+  }
+);
+
+// Cached unsupported adaptors are not downloaded again
+test.serial(
+  `openfn metadata -S "${state}" -a openfn@3.0.0 --log info --log-json `,
+  async (t) => {
+    const { stdout } = await run(t.title);
+
+    t.regex(stdout, /known to not support metadata \(cached\)/);
+    t.regex(stdout, /No metadata helper found/);
+    // Should NOT try to install or remove again
+    t.notRegex(stdout, /Installing packages/);
+    t.notRegex(stdout, /Removing unsupported adaptor from disk/);
+    t.notRegex(stdout, /successfully removed @openfn\/language-openfn_/);
+  }
+);
+
+// metadata should work with implicit @latest (note that it'll still autoinstall)
+test.serial(
+  `openfn metadata -S "${state}" -a openfn --log-json --log info`,
+  async (t) => {
+    const { stdout } = await run(t.title);
+    t.log(stdout);
+
+    t.regex(stdout, /Generating metadata/);
+    t.regex(stdout, /Installing packages.../);
+    t.regex(stdout, /Installed @openfn\/language-openfn@/);
+    t.regex(stdout, /Installation complete in \d+\.\d+s/);
+    t.regex(stdout, /No metadata helper found/);
+    t.regex(stdout, /Removing unsupported adaptor from disk/);
+    t.regex(stdout, /Adaptor removed and marked as unsupported/);
+    t.regex(stdout, /successfully removed @openfn\/language-openfn_/i);
+  }
+);
+
+// ignore cache with --keep-unsupported
+test.serial(
+  `openfn metadata -S "${state}" -a http@7.0.0 --keep-unsupported --log-json --log info`,
+  async (t) => {
+    const { stdout } = await run(t.title);
+    t.regex(stdout, /Generating metadata/);
+    t.regex(stdout, /Installing packages.../);
+    t.regex(stdout, /Installed @openfn\/language-http@/);
+    t.regex(stdout, /Installation complete in \d+\.\d+s/);
+    t.regex(stdout, /No metadata helper found/);
+    t.regex(
+      stdout,
+      /Keeping unsupported adaptor as requested by --keep-unsupported flag/
+    );
+    // Should NOT try to remove the package
+    t.notRegex(stdout, /Removing unsupported adaptor from disk/);
+  }
+);
+
+// This test validates that even with --keep-unsupported, the cache works
+test.serial(
+  `openfn metadata -fS "${state}" -a http@7.0.0 --keep-unsupported --log-json --log info`,
+  async (t) => {
+    const { stdout } = await run(t.title);
+
+    t.regex(stdout, /known to not support metadata/);
+    t.regex(stdout, /No metadata helper found/);
+    // Should NOT try to install again
+    t.notRegex(stdout, /Installing packages/);
   }
 );