+++ packages/ws-worker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ws-worker
 
+## 1.14.2
+
+### Patch Changes
+
+- Updated dependencies [f04acef]
+  - @openfn/runtime@1.7.1
+  - @openfn/engine-multi@1.6.8
+
 ## 1.14.1
 
 ### Patch Changes